+++ src/preload/index.d.ts
@@ -28,6 +28,7 @@ export interface ApplicationAPI {
   stopProject: (baseDir: string) => void;
   restartProject: (baseDir: string) => void;
   runPrompt: (baseDir: string, prompt: string, mode?: Mode) => void;
+  redoLastUserPrompt: (baseDir: string, mode: Mode) => void;
   answerQuestion: (baseDir: string, answer: string) => void;
   loadInputHistory: (baseDir: string) => Promise<string[]>;
   dialog: {