+++ src/preload/index.ts
@@ -49,6 +49,7 @@ const api: ApplicationAPI = {
   stopProject: (baseDir) => ipcRenderer.send('stop-project', baseDir),
   restartProject: (baseDir) => ipcRenderer.send('restart-project', baseDir),
   runPrompt: (baseDir, prompt, mode) => ipcRenderer.send('run-prompt', baseDir, prompt, mode),
+  redoLastUserPrompt: (baseDir, mode) => ipcRenderer.send('redo-last-user-prompt', baseDir, mode),
   answerQuestion: (baseDir, answer) => ipcRenderer.send('answer-question', baseDir, answer),
   loadInputHistory: (baseDir) => ipcRenderer.invoke('load-input-history', baseDir),
   dialog: {