+++ src/renderer/src/components/message/MessageBar.tsx
@@ -2,20 +2,22 @@ import { useRef, useState } from 'react';
 import { FaArrowDown, FaArrowUp, FaDollarSign, FaEllipsisVertical } from 'react-icons/fa6';
 import { useTranslation } from 'react-i18next';
 import { UsageReportData } from '@common/types';
-import { MdDeleteForever } from 'react-icons/md';
+import { MdDeleteForever, MdRedo } from 'react-icons/md';
 
-import { useClickOutside } from '../../hooks/useClickOutside';
 import { IconButton } from '../common/IconButton';
 
 import { CopyMessageButton } from './CopyMessageButton';
 
+import { useClickOutside } from '@/hooks/useClickOutside';
+
 type Props = {
   content: string;
   usageReport?: UsageReportData;
-  removeMessage?: () => void;
+  remove?: () => void;
+  redo?: () => void;
 };
 
-export const MessageBar = ({ content, usageReport, removeMessage }: Props) => {
+export const MessageBar = ({ content, usageReport, remove, redo }: Props) => {
   const { t } = useTranslation();
   const [isMenuOpen, setIsMenuOpen] = useState(false);
   const menuRef = useRef<HTMLDivElement>(null);
@@ -26,7 +28,12 @@ export const MessageBar = ({ content, usageReport, removeMessage }: Props) => {
   });
 
   const handleRemoveClick = () => {
-    removeMessage?.();
+    remove?.();
+    setIsMenuOpen(false);
+  };
+
+  const handleRedoClick = () => {
+    redo?.();
     setIsMenuOpen(false);
   };
 
@@ -48,7 +55,7 @@ export const MessageBar = ({ content, usageReport, removeMessage }: Props) => {
         </div>
       )}
       <CopyMessageButton content={content} className="transition-colors text-neutral-700 hover:text-neutral-100" alwaysShow={true} />
-      {removeMessage && (
+      {(remove || redo) && (
         <div ref={buttonRef}>
           <IconButton
             icon={<FaEllipsisVertical className="w-4 h-4" />}
@@ -57,16 +64,27 @@ export const MessageBar = ({ content, usageReport, removeMessage }: Props) => {
           />
         </div>
       )}
-      {isMenuOpen && removeMessage && (
+      {isMenuOpen && (remove || redo) && (
         <div ref={menuRef} className="absolute right-0 bottom-full  w-[120px] bg-neutral-800 border border-neutral-700 rounded shadow-lg z-10">
           <ul>
-            <li
-              className="flex items-center gap-1 px-2 py-1 text-xxs text-neutral-100 hover:bg-neutral-700 cursor-pointer transition-colors"
-              onClick={handleRemoveClick}
-            >
-              <MdDeleteForever className="w-4 h-4" />
-              <span className="whitespace-nowrap mb-[-4px]">{t('message.delete')}</span>
-            </li>
+            {redo && (
+              <li
+                className="flex items-center gap-1 px-2 py-1 text-xxs text-neutral-100 hover:bg-neutral-700 cursor-pointer transition-colors"
+                onClick={handleRedoClick}
+              >
+                <MdRedo className="w-4 h-4" />
+                <span className="whitespace-nowrap mb-[-4px]">{t('messages.redo')}</span>
+              </li>
+            )}
+            {remove && (
+              <li
+                className="flex items-center gap-1 px-2 py-1 text-xxs text-neutral-100 hover:bg-neutral-700 cursor-pointer transition-colors"
+                onClick={handleRemoveClick}
+              >
+                <MdDeleteForever className="w-4 h-4" />
+                <span className="whitespace-nowrap mb-[-4px]">{t('messages.delete')}</span>
+              </li>
+            )}
           </ul>
         </div>
       )}