+++ src/renderer/src/components/message/UserMessageBlock.tsx
@@ -12,9 +12,10 @@ type Props = {
   allFiles: string[];
   renderMarkdown: boolean;
   onRemove?: () => void;
+  redo?: () => void;
 };
 
-export const UserMessageBlock = ({ baseDir, message, allFiles, renderMarkdown, onRemove }: Props) => {
+export const UserMessageBlock = ({ baseDir, message, allFiles, renderMarkdown, onRemove, redo }: Props) => {
   const baseClasses = 'rounded-md p-3 mb-2 max-w-full text-xs bg-neutral-850 border border-neutral-800 text-gray-100';
 
   return (
@@ -25,7 +26,7 @@ export const UserMessageBlock = ({ baseDir, message, allFiles, renderMarkdown, o
         </div>
         <div className="flex-grow-1 w-full overflow-hidden">{parseMessageContent(baseDir, message.content, allFiles, renderMarkdown)}</div>
       </div>
-      <MessageBar content={message.content} removeMessage={onRemove} />
+      <MessageBar content={message.content} remove={onRemove} redo={redo} />
     </div>
   );
 };